WIP: Use a separate node for permissions.
