+++ src/article/converter/r2t/jats2internal.js
@@ -34,7 +34,7 @@ import createJatsImporter from './createJatsImporter'
       isbn?,            // -> article-record
       (((fpage,lpage?)?,page-range?)|elocation-id)?,  // -> article-record
       history?,         // -> article-record
-      permissions?,
+      permissions?,     // -> article-record
       self-uri*,        // not supported yet
       (related-article,related-object)*, // not supported yet
       abstract?,        // -> content.abstract
@@ -66,7 +66,7 @@ export default function jats2internal (jats) {
   _populateAuthors(doc, jats, jatsImporter)
   _populateEditors(doc, jats, jatsImporter)
   _populateAwards(doc, jats)
-  _populateArticleRecord(doc, jats)
+  _populateArticleRecord(doc, jats, jatsImporter)
   _populateKeywords(doc, jats)
   _populateSubjects(doc, jats)
 
@@ -218,16 +218,19 @@ function _populateAwards (doc, jats) {
   })
 }
 
-function _populateArticleRecord (doc, jats) {
+function _populateArticleRecord (doc, jats, jatsImporter) {
   let articleMetaEl = jats.find('article > front > article-meta')
+  // const permissionsEl = articleMetaEl.find('permissions')
   let articleRecord = doc.get('article-record')
+
   Object.assign(articleRecord, {
     elocationId: getText(articleMetaEl, 'elocation-id'),
     fpage: getText(articleMetaEl, 'fpage'),
     lpage: getText(articleMetaEl, 'lpage'),
     issue: getText(articleMetaEl, 'issue'),
     volume: getText(articleMetaEl, 'volume'),
-    pageRange: getText(articleMetaEl, 'page-range')
+    pageRange: getText(articleMetaEl, 'page-range'),
+    // permission: jatsImporter.convertElement(permissionsEl).id
   })
   const articleDateEls = articleMetaEl.findAll('history > date, pub-date')
   articleDateEls.forEach(dateEl => {
